edit: Fix InvalidDistanceMatrix error for valid matrices
Fixes issue #2.

The bug occured because the code that fetched the distances
associated with a substitute/edit/insert operation used an Option type
to denote the distances, and set it to None if the operation wasn't
valid from the current position in the distance matrix. Later, the
std::cmp::min function was used to find the minimum distance. However,
reasonably, the min function considers the None variant of the Option
enum to be less than any value encapsulated in a Some variant. So e.g. a
None value would be less than Some(0). This lead to the error path that
threw the InvalidDistanceMatrix error.

The fix was to simply not use an Option type here, and instead to use
usize::MAX as a sentinel value for operations that weren't valid from
the current position in the distance matrix.

Also adds a test adapted from the bug report in issue #2.